Add extra coordinates to atoms at VDW radii
Adds additional points to each atom in a sphere, such that points are
evenly distributed.  This is done based on the voxel-size being used.

In this way, asking for higher resolution does not lead to strange
tunnels etc. being created and the number of e.g. pores being computed
remains roughly constant.