Fix biological assemblies
Biological assemblies are recorded as being individual models for each
unit in the assembly.  This was causing problems with the way the PDB
was cleaned by bio-python.

Now models are left in (NMR structures only have one model when
downloaded as a biological assembly).

Additionally, in revamping the bio-python selection class a change was
made to allow non-protein residues, such as would be the case when
analyzing a designed RNA structure.

Furthermore, to support annotations on these structures with multiple
models, the way the annotation was written out was improved to work
across multiple models (at least in Pymol).