Implement subsetting of exposed voxels
Previously when the type of buried voxel group was being determined a
search was performed over ALL exposed voxels to see which of the buried
voxels were neighboring an exposed voxel and were therefore part of the
surface.

Now, the exposed voxels are first subset to ONLY those that themselves
neighbor a buried voxel, thereby dramatically speeding up the assignment
of voxel group type.

The python version of the code actually results in a net-slow down
because the cost to subset the exposed voxels is relatively high, but
the C version of that function leads to a considerable speedup.

Unit tests also accompany this commit.