Script to filter a list of PDBs by size
Allows filtering a list of PDB IDs (one per line) into a smaller list by
setting min and max cutoffs on:

- number of atoms
- number of residues
- number of chains

Numbers are for the biological assembly.

Metrics are retrieved from the RCSB (only works on PDBs with an RCSB PDB
ID not on custom files), and are saved locally.